test(worker): isolate artifact session scenario
